requirements.txt
# Core
python-dotenv>=1.0.0
httpx>=0.24.0

# ETL & Data Processing
pydantic>=2.0.0
pandas>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Testing
pytest>=7.0.0

# (опционально, для безопасной парсировки дат)
python-dateutil>=2.8.0